+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 CasperJS Changelog
 ==================
 
+XXXX-XX-XX, v1.0.0
+------------------
+
+This version is yet to be released.
+
+- fixed [#259](https://github.com/n1k0/casperjs/issues/259) - enhanced `Tester.assertField()` method, which can now tests for other field types than `input`s.
+- added [`ClientUtils.getFieldValue()`](http://casperjs.org/api.html#clientutils.getFieldValue)
+
 2012-10-23, v1.0.0-RC3
 ----------------------